[Fix] separate Request from Request_parser
