+++ includes/Request_parser.hpp
@@ -7,23 +7,7 @@
 #include <unistd.h>
 #include <sys/stat.h>
 #include <sys/types.h>
-#include "Server.hpp"
-
-
-struct Request
-{
-	int				method;	
-	std::string		url; // original whole string
-	std::string		path; // url without query
-	Location		*location;
-	bool			is_cgi;
-	bool			is_directory;
-	std::string		connection;
-	int				content_length;
-	std::string		content_type;
-	std::string		body;
-	int				status_code; // 초기값 0 
-};
+#include "Client.hpp"
 
 class Request_parser
 {
@@ -44,7 +28,7 @@ class Request_parser
 		~Request_parser();
 		Request								request;
 
-		void		parse_request(Server *server);
+		void		parse_request(Client& client);
 };
 
 #endif
